requirements.txt
django==1.10
djangorestframework>=3.5.3
django-cors-headers>=2.0.2
requests>=2.13
djangorestframework-jwt>=1.9.0

